requirements.txt
requests
python-dotenv
openai
google-generativeai
telegramify-markdown